Spelling out states in SocCaus-07 +Appr-12 fix
This required changes to codes, parameters, values, and SocCaus.md. In addition I fixed headers for SocCaus-02 in SocCaus.md.

Appr-12 changes from the previous check-in required that the values in addition receive ()s, e.g. A (apprehensive). This was fixed in the values.csv and also necessitated a change to appr.R to facilitate the derivations of Appr-12a & 12b.